Texture class can be used by any Tiler
Now gives image path (or image bytes) as parameter of Texture constructor. Arguments/calls specific to the CityTiler are moved in the CityTiler classes
